requirements.txt
django-celery
mutagen
django